custom-elements/registering-custom-elements/definition-construction-algorithm-test-002.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element local name should be lowercased</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If DOCUMENT is an HTML document, convert NAME to lowercase">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

function doTest(element) {
    var doc = newHTMLDocument();
    
    var obj = doc.createElement(element);
    
    var name = 'x-foo-' + element;
    
    var proto = Object.create(obj.constructor.prototype);
    
    var XFoo = doc.register(name, {prototype: proto, extends: element.toUpperCase()});
    
    var xfoo = new XFoo();
        
    assert_equals(xfoo.localName, element, 'Local name should be lowercased');    
}


var parameters = [];

for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
    if (HTML5_DOCUMENT_ELEMENTS.indexOf(HTML5_ELEMENTS[i]) == -1) {
        parameters[i] = ['Custom element type ' + HTML5_ELEMENTS[i], HTML5_ELEMENTS[i]];
    }
}

generate_tests(doTest, parameters, 
        'Custom element local name should be lowercased. Test constructor');

</script>
</body>
</html>
